🚀 Added New Article, Enhanced Website, Open-Sourced Project, Social Links, Footers, Privacy & Terms! 📝🌐
In this update, we've been hard at work to bring you exciting improvements to Beaver Notes! We added a new article to share insights and tips, polished the website with a fresh design, and made the source code public for transparency and community collaboration. Get ready for enhanced features, social links, footers, and improved privacy and terms to create a more seamless note-taking experience. Your feedback matters to us, so join us in shaping the future of note-taking! 🎉